chore: Merge useContractExchange with useContract
