+++ src/hooks/ENS/useENSAddress.ts
@@ -2,7 +2,7 @@ import { namehash } from 'ethers/lib/utils'
 import { useMemo } from 'react'
 import { useSingleCallResult } from '../../state/multicall/hooks'
 import isZero from '../../utils/isZero'
-import { useENSRegistrarContract, useENSResolverContract } from '../useContractExchange'
+import { useENSRegistrarContract, useENSResolverContract } from '../useContract'
 import useDebounce from '../useDebounce'
 
 /**
