+++ src/hooks/ENS/useENSName.ts
@@ -3,7 +3,7 @@ import { useMemo } from 'react'
 import { useSingleCallResult } from '../../state/multicall/hooks'
 import { isAddress } from '../../utils'
 import isZero from '../../utils/isZero'
-import { useENSRegistrarContract, useENSResolverContract } from '../useContractExchange'
+import { useENSRegistrarContract, useENSResolverContract } from '../useContract'
 import useDebounce from '../useDebounce'
 
 /**
