+++ src/hooks/useCurrentBlockTimestamp.ts
@@ -1,6 +1,6 @@
 import { BigNumber } from 'ethers'
 import { useSingleCallResult } from '../state/multicall/hooks'
-import { useMulticallContract } from './useContractExchange'
+import { useMulticallContract } from './useContract'
 
 // gets the current timestamp from the blockchain
 export default function useCurrentBlockTimestamp(): BigNumber | undefined {
