+++ src/hooks/useTotalSupply.ts
@@ -1,6 +1,6 @@
 import { BigNumber } from '@ethersproject/bignumber'
 import { Token, TokenAmount } from '@pancakeswap/sdk'
-import { useTokenContract } from './useContractExchange'
+import { useTokenContract } from './useContract'
 import { useSingleCallResult } from '../state/multicall/hooks'
 
 // returns undefined if input token is undefined, or fails to get token contract,
