+++ src/hooks/useWrapCallback.ts
@@ -4,7 +4,7 @@ import { tryParseAmount } from '../state/swap/hooks'
 import { useTransactionAdder } from '../state/transactions/hooks'
 import { useCurrencyBalance } from '../state/wallet/hooks'
 import { useActiveWeb3React } from './Auth'
-import { useWETHContract } from './useContractExchange'
+import { useWETHContract } from './useContract'
 
 export enum WrapType {
   NOT_APPLICABLE,
